moved messaging for managability, fixed version disconnect
fixed messaging for managability, fixed version disconnect so that any
bad nodes will disconnect after april first
