+++ src/main.h
@@ -17,10 +17,12 @@
 #include <list>
 
 class CWallet;
+class CWalletTx;
 class CBlock;
 class CBlockIndex;
 class CKeyItem;
 class CReserveKey;
+class COutPoint;
 
 class CAddress;
 class CInv;
@@ -59,7 +61,12 @@ extern CScript COINBASE_FLAGS;
 
 
 
-
+extern CMedianFilter<int> cPeerBlockCounts;
+extern std::map<uint256, CBlock*> mapOrphanBlocks;
+extern std::multimap<uint256, CBlock*> mapOrphanBlocksByPrev;
+extern std::map<uint256, CDataStream*> mapOrphanTransactions;
+extern std::map<uint256, std::map<uint256, CDataStream*> > mapOrphanTransactionsByPrev;
+extern CScript COINBASE_FLAGS;
 extern CCriticalSection cs_main;
 extern std::map<uint256, CBlockIndex*> mapBlockIndex;
 extern uint256 hashGenesisBlock;
@@ -115,6 +122,14 @@ int GetNumBlocksOfPeers();
 bool IsInitialBlockDownload();
 std::string GetWarnings(std::string strFor);
 bool GetTransaction(const uint256 &hash, CTransaction &tx, uint256 &hashBlock);
+
+void Inventory(const uint256& hash);
+uint256 GetOrphanRoot(const CBlock* pblock);
+bool GetTransaction(const uint256& hashTx, CWalletTx& wtx);
+unsigned int LimitOrphanTxSize(unsigned int nMaxOrphans);
+bool AddOrphanTx(const CDataStream& vMsg);
+void EraseOrphanTx(uint256 hash);
+void ResendWalletTransactions();
 
 
 
@@ -1612,6 +1627,10 @@ public:
      */
     static CAlert getAlertByHash(const uint256 &hash);
 };
+
+extern std::map<uint256, CAlert> mapAlerts;
+extern CCriticalSection cs_mapAlerts;
+
 
 class CTxMemPool
 {
